Added init logic to FactionSystem + data parsing. Also fixed SpriteComponent constructor and SpriteOutlineComponent warning

+++ src/ECS/Components/SpriteOutlineComponent.cpp
@@ -49,7 +49,10 @@ void SpriteOutlineComponent::draw()
 	const auto scale = transformComponent->getScale();
 	for (const auto& outlinePosition : outlinePositions)
 	{
-		tempDestRect = SDL_Rect(destRect.x + outlinePosition.x * scale, destRect.y + outlinePosition.y * scale, destRect.w, destRect.h);
+		tempDestRect = SDL_Rect(
+			destRect.x + static_cast<int>(outlinePosition.x * scale),
+			destRect.y + static_cast<int>(outlinePosition.y * scale),
+			destRect.w, destRect.h);
 		TextureManager::draw(texture, srcRect, tempDestRect, spriteFlip);
 	}
 }
